feat: Update SafetyPerCellProcessor to use dynamic safety field
The code changes in `safety_polygon_processor.py` modify the `SafetyPerCellProcessor` class to include a new parameter `safety_field`. This parameter allows the user to specify the field name that contains the safety values in the input layer.

This change enables greater flexibility in processing safety data by allowing different field names to be used.